feat(client): Refactoring
