+++ src/new-client/src/App/Header/PWA/PWAInstallModal.styles.ts
@@ -1,0 +1,28 @@
+import styled from "styled-components"
+
+export const MainTitle = styled.div`
+  font-size: 1.19rem;
+  line-height: 2rem;
+`
+
+export const Text = styled.div`
+  font-size: 0.81rem;
+  color: rgb(187, 187, 187);
+`
+
+export const DeviceText = styled.div`
+  margin: 10px 0 30px;
+  font-size: 0.81rem;
+  color: rgb(187, 187, 187);
+  line-height: 2px;
+`
+
+export const ModalWrapper = styled.div`
+  text-align: center;
+`
+
+export const Icon = styled.div`
+  display: inline-flex;
+  vertical-align: middle;
+  margin-bottom: 4px;
+`
